add plotting facades to DataSet and remove the shallow ones from DataSet_Plot.py
